Refactor: extract common list index and section logic from steps and ingredients lists
