+++ src/components/organisms/recipe/detail/stepsList/index.jsx
@@ -1,60 +1,19 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
-import NoItems from '../../../list/noItems.jsx';
+import List from '../../../list/index.jsx';
 
 import Section from './components/section.jsx';
 import Item from './components/item.jsx';
 
 const Component = ({ className, steps }) => {
-  let rowCounter = -1;
-
-  const [checked, setChecked] = useState([]);
-
-  const handleToggle = value => () => {
-    const currentIndex = checked.indexOf(value);
-    const newChecked = [...checked];
-
-    if (currentIndex === -1) {
-      newChecked.push(value);
-    } else {
-      newChecked.splice(currentIndex, 1);
-    }
-
-    setChecked(newChecked);
-  };
-
-  return steps && steps.length > 0 ? (
-    steps.map((step, index) => (
-      <Section
-        {...step}
-        index={index}
-        rowCounter={rowCounter}
-        key={`${index}-${rowCounter}`}
-        className={className}
-        renderItem={(item, itemIndex) => {
-          rowCounter += 1;
-
-          return (
-            item && (
-              <Item
-                index={itemIndex}
-                rowCounter={rowCounter}
-                step={item}
-                key={`${itemIndex}-${rowCounter}`}
-                value={checked.indexOf(rowCounter) !== -1}
-                handleClick={handleToggle(rowCounter)}
-              />
-            )
-          );
-        }}
-      />
-    ))
-  ) : (
-      <NoItems>
-        No steps required
-      </NoItems>
-    );
+  return <List
+    className={className}
+    items={steps}
+    noItemsMsg='No steps required'
+    Section={(props) => <Section {...props} />}
+    Item={(props) => <Item {...props} />}
+  />;
 };
 
 Component.defaultProps = {
